Adding Playwright tests for Accordion
