+++ packages/web-components/src/components/rux-accordion/rux-accordion.scss
@@ -4,7 +4,7 @@
 
 :host([disabled]) {
     cursor: not-allowed;
-    opacity: var(--disabled-opacity);
+    opacity: var(--opacity-disabled);
 }
 :host([disabled]) ::slotted(*) {
     pointer-events: none;
