+++ packages/web-components/tsconfig.json
@@ -16,6 +16,11 @@
         "typeRoots": ["node_modules/@types"],
         "resolveJsonModule": true
     },
-    "include": ["src", "src/**/*.ts", "tests/**/*.ts"],
+    "include": [
+        "src",
+        "src/**/*.ts",
+        "tests/**/*.ts",
+        "tests/accordion.spec.tsx"
+    ],
     "exclude": ["node_modules", "src/stories"]
 }
